- a few minor bug fixes related to code warnings - code cleanup based on additional gcc warnings - enabled additional gcc compiler warnings

+++ source/glest_game/ai/ai.cpp
@@ -717,7 +717,7 @@ void Ai::addExpansion(const Vec2i &pos) {
 	expansionPositions.push_front(pos);
 
 	//remove expansion if queue is list is full
-	if(expansionPositions.size() > maxExpansions){
+	if((int)expansionPositions.size() > maxExpansions){
 		expansionPositions.pop_back();
 	}
 }
