+++ source/glest_game/game/console.cpp
@@ -108,12 +108,12 @@ void Console::addLine(string line, bool playSound, int playerIndex, Vec3f textCo
 			storedLines.clear();
 		}
 		lines.insert(lines.begin(), info);
-		if(lines.size() > maxLines) {
+		if((int)lines.size() > maxLines) {
 			lines.pop_back();
 		}
 		if(onlyChatMessagesInStoredLines==false || info.PlayerIndex!=-1) {
 			storedLines.insert(storedLines.begin(), info);
-			if(storedLines.size() > maxStoredLines) {
+			if((int)storedLines.size() > maxStoredLines) {
 				storedLines.pop_back();
 			}
 		}
@@ -144,11 +144,11 @@ void Console::addLine(string line, bool playSound, string playerName, Vec3f text
 		//printf("info.PlayerIndex = %d, line [%s]\n",info.PlayerIndex,info.originalPlayerName.c_str());
 
 		lines.insert(lines.begin(), info);
-		if(lines.size() > maxLines) {
+		if((int)lines.size() > maxLines) {
 			lines.pop_back();
 		}
 		storedLines.insert(storedLines.begin(), info);
-		if(storedLines.size() > maxStoredLines) {
+		if((int)storedLines.size() > maxStoredLines) {
 			storedLines.pop_back();
 		}
 	}
@@ -181,25 +181,25 @@ bool Console::isEmpty() {
 }
 
 string Console::getLine(int i) const {
-	if(i < 0 || i >= lines.size())
+	if(i < 0 || i >= (int)lines.size())
 		throw megaglest_runtime_error("i >= Lines.size()");
 	return lines[i].text;
 }
 
 string Console::getStoredLine(int i) const {
-	if(i < 0 || i >= storedLines.size())
+	if(i < 0 || i >= (int)storedLines.size())
 		throw megaglest_runtime_error("i >= storedLines.size()");
 	return storedLines[i].text;
 }
 
 ConsoleLineInfo Console::getLineItem(int i) const {
-	if(i < 0 || i >= lines.size())
+	if(i < 0 || i >= (int)lines.size())
 		throw megaglest_runtime_error("i >= Lines.size()");
 	return lines[i];
 }
 
 ConsoleLineInfo Console::getStoredLineItem(int i) const {
-	if(i < 0 || i >= storedLines.size())
+	if(i < 0 || i >= (int)storedLines.size())
 		throw megaglest_runtime_error("i >= storedLines.size()");
 	return storedLines[i];
 }
